[BACKLOG-13089] Mavenize pentaho-kettle/plugins/kettle-dummy-plugin
