+++ build.properties
@@ -80,7 +80,9 @@ coverage.html.dir=${coverage.dir}/html
 
 includeantruntime=false
 
-plugin.list=kettle-gpload-plugin,kettle-palo-plugin,kettle-openerp-plugin,kettle5-log4j-plugin,googleanalytics,ms-access-bulk-loader,lucid-db-streaming-loader,xml-input,xml-input-stream,ssh2-plugin,lucid-db-bulk-loader,salesforce,kettle-dummy-plugin,kettle-s3csvinput-plugin,kettle-shapefilereader-plugin,kettle-version-checker,pdi-pur-plugin,kettle-xml-plugin,kettle-json-plugin,kettle-sap-plugin
+
+plugin.list=kettle-gpload-plugin,kettle-palo-plugin,kettle-openerp-plugin,kettle5-log4j-plugin,googleanalytics,ms-access-bulk-loader,lucid-db-streaming-loader,xml-input,xml-input-stream,ssh2-plugin,lucid-db-bulk-loader,salesforce,kettle-s3csvinput-plugin,kettle-shapefilereader-plugin,kettle-version-checker,pdi-pur-plugin,kettle-xml-plugin,kettle-json-plugin,kettle-sap-plugin
+
 module.list=core,engine,dbdialog,ui
 all-subprojects.list=${module.list},${plugin.list}
 assembly.stage.dir=${assembly.dir}/bin/stage
