+++ plugins/kettle-dummy-plugin/.gitignore
@@ -1,0 +1,4 @@
+.idea
+target
+bin
+*.iml
