+++ plugins/kettle-dummy-plugin/assemblies/plugin/pom.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.pentaho.kettle.plugins</groupId>
+    <artifactId>kettle-dummy-plugin-assemblies</artifactId>
+    <version>8.0-SNAPSHOT</version>
+  </parent>
+  <groupId>org.pentaho.kettle.plugins</groupId>
+  <artifactId>kettle-dummy-plugin</artifactId>
+  <version>8.0-SNAPSHOT</version>
+  <packaging>pom</packaging>
+  <name>${project.artifactId}</name>
+  <description></description>
+  <dependencies>
+    <dependency>
+      <groupId>org.pentaho.kettle.plugins</groupId>
+      <artifactId>kettle-dummy-plugin-core</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+  </dependencies>
+</project>
