--- plugins/kettle-dummy-plugin/build.properties
@@ -1,11 +1,0 @@
-project.revision=8.0-SNAPSHOT
-ivy.artifact.group=pentaho-kettle
-ivy.artifact.id=kettle-dummy-plugin
-impl.vendor=Pentaho Corporation
-impl.title=Dummy/Example Plugin
-impl.productID=${ivy.artifact.id}
-dev-lib.dir=dev-lib
-
-dependency.kettle.revision=8.0-SNAPSHOT
-
-javadoc.packagenames=org.pentaho.*,be.ibridge.kettle.*,pdi.jobentry.dummy.*
