+++ plugins/kettle-dummy-plugin/core/src/main/filters/version.properties
@@ -1,0 +1,5 @@
+#Dummy/Example Plugin build information
+#Tue, 02 May 2017 09:49:31 +0100
+
+version=8.0-SNAPSHOT.development
+builddate=2017/05/02 09\:49
