+++ plugins/kettle-dummy-plugin/core/src/main/java/org/pentaho/di/be/ibridge/kettle/dummy/DummyPluginData.java
@@ -1,0 +1,41 @@
+/*! ******************************************************************************
+ *
+ * Pentaho Data Integration
+ *
+ * Copyright (C) 2002-2017 by Pentaho : http://www.pentaho.com
+ *
+ *******************************************************************************
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ ******************************************************************************/
+
+package org.pentaho.di.be.ibridge.kettle.dummy;
+
+import org.pentaho.di.core.row.RowMetaInterface;
+import org.pentaho.di.trans.step.BaseStepData;
+import org.pentaho.di.trans.step.StepDataInterface;
+
+/**
+ * 
+ * 
+ * @author Matt
+ * @since  24-mrt-2005
+ */
+public class DummyPluginData extends BaseStepData implements StepDataInterface {
+  public RowMetaInterface outputRowMeta;
+
+  public DummyPluginData() {
+    super();
+  }
+}
