+++ plugins/kettle-dummy-plugin/core/src/main/resources/DPL.png
