+++ plugins/kettle-dummy-plugin/core/src/main/resources/org/pentaho/di/be/ibridge/kettle/dummy/messages/messages_en_US.properties
@@ -1,0 +1,8 @@
+DummyPluginDialog.Shell.Title=Dummy Plugin test
+DummyPluginDialog.StepName.Label=Step name 
+DummyPluginDialog.ValueName.Label=Value name 
+DummyPluginDialog.ValueToAdd.Label=Value to add :
+DummyPluginDialog.Button.Edit.Label=&Edit...
+DummyPlugin.Step.Name=Example Step
+DummyPlugin.Step.Description=This is a plugin example step
+Deprecated=Deprecated
