+++ plugins/kettle-dummy-plugin/core/src/main/resources/org/pentaho/di/pdi/jobentry/dummy/messages/messages_en_US.properties
@@ -1,0 +1,4 @@
+DummyPluginDialog.Shell.Title=Dummy Plugin test
+DummyPlugin.Job.Name=Example Job
+DummyPlugin.Job.Description=This is an example job
+Deprecated=Deprecated
