+++ plugins/kettle-dummy-plugin/core/src/test/java/org/pentaho/di/pdi/jobentry/dummy/TestDummyJob.java
@@ -1,0 +1,37 @@
+/*! ******************************************************************************
+ *
+ * Pentaho Data Integration
+ *
+ * Copyright (C) 2002-2016 by Pentaho : http://www.pentaho.com
+ *
+ *******************************************************************************
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ ******************************************************************************/
+
+/*
+ * Created on 17-Feb-07
+ * Actualis Center
+ *
+ */
+package org.pentaho.di.pdi.jobentry.dummy;
+
+public class TestDummyJob {
+
+  // @Test
+  public void testDummyJob() throws Exception {
+    DummyJob sp = new DummyJob( "testdata", "testout", ".*.xml" );
+    sp.process();
+  }
+}
