--- plugins/kettle-dummy-plugin/ivy.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ivy-module version="2.0" xmlns:m="http://ant.apache.org/ivy/maven">
-  <info organisation="${ivy.artifact.group}" module="${ivy.artifact.id}" revision="${dependency.kettle.revision}" status="integration" publication="20130425215250"/>
-
-  <configurations>
-    <conf name="default"/>
-    <conf name="dev"/>
-    <conf name="test" visibility="private"/>
-    <conf name="source"/>
-    <conf name="zip"/>
-  </configurations>
-
-  <publications>
-    <artifact name="${ivy.artifact.id}" type="jar" conf="dev"/>
-    <artifact name="${ivy.artifact.id}" type="zip" conf="zip" />
-    <artifact name="${ivy.artifact.id}" m:classifier="sources" type="source" ext="jar" conf="source"/>
-    <artifact name="${ivy.artifact.id}" m:classifier="sources" type="source" ext="zip" conf="source"/>
-  </publications>
-
-  <dependencies defaultconf="dev->default">
-    <dependency org="pentaho-kettle" name="kettle-core" rev="${dependency.kettle.revision}" changing="true"/>  
-    <dependency org="pentaho-kettle" name="kettle-engine" rev="${dependency.kettle.revision}" changing="true"/>
-    <dependency org="pentaho-kettle" name="kettle-ui-swt" rev="${dependency.kettle.revision}" changing="true"/>
-    
-    <!-- SWT it required to compile any version of any architecture will work -->
-    <dependency org="org.eclipse.swt" name="org.eclipse.swt.gtk.linux.x86_64" rev="4.6" transitive="false"/>
-    <dependency org="org.eclipse" name="jface" rev="3.3.0-I20070606-0010" transitive="false"/> 
-    
-    <dependency org="junit" name="junit" rev="4.7" conf="test->default"/> 
-    
-  </dependencies>
-</ivy-module>
